--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/ControllerHandlerAdapter.java
@@ -1,21 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.asis;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.HandlerAdapter;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        Controller controller = (Controller) handler;
-        String viewName = controller.execute(request, response);
-        return new ModelAndView(viewName);
-    }
-}
